Removed duplicate constant aliasing
The same gets done in dm-serializer/common.rb which
gets required by all the various to_* serializers.

In general, it might not even be necessary to do
any of these aliasings at all in plugin land, since
dm-core actually does it itself already, and needs
to be required before the plugins do their thing
anyway (most of the times at least).

I'm not sure if this change will actually eliminate
all const redefinition warnings, but executing this
code in that place is definitely not necessary.

[#1302]